A program for multiplying polynomials by the Karatsuba algorithm
This is a program for multiplying large polynomials by the karatsuba algorithm. Unfortunately, this version only works for those polynomials where the number of terms is a power of two.

The algorithm itself is in the file KaratsubaMul.c and KaratsubaMul.h as header.

The remaining files are created for testing the algorithm.

In order to add your test, create your own file with test data in the Tests folder and write its name to the file MainTests.txt . After that, just type 'make run' in the command line. If your program is not compiled, type 'make' commandbefore that.